Exercises 4.15 - 4.23
Exercises 4.15 - 4.23

+++ Part4/bloglist/controllers/blogs.js
@@ -1,26 +1,36 @@
 const blogsRouter = require("express").Router()  
-const Blog = require("../models/blog")  
+const Blog = require("../models/blog")
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
+
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs.map(blog => blog.toJSON()))
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
+  response.json(blogs)
 })  
 
-blogsRouter.post("/", async (request, response, next) => {
-  const body = request.body  
-
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes? body.likes : 0
-  })
-
-  blog.save()
-    .then(savedBlog => {
-      response.status(201).json(savedBlog)
+blogsRouter.post("/", middleware.userExtractor, async (request, response, next) => {
+  User.findById(request.user._id).then((user) => {
+    const blog = new Blog({
+      title: request.body.title,
+      author: request.body.author,
+      url: request.body.url,
+      likes: request.body.likes? request.body.likes : 0,
+      user: request.user
     })
-    .catch(error => next(error))
+
+    blog.save()
+      .then(savedBlog => {
+        user.blogs = user.blogs.concat(savedBlog._id)
+        user.save().then(() => {
+          response.status(201).json(savedBlog)
+        }).catch(error => next(error))
+      })
+      .catch(error => next(error))
+  }).catch(error => next(error))
+  
 })  
 
 blogsRouter.get('/:id', async (request, response) => {
@@ -32,12 +42,16 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() === request.user._id.toString()){
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+  }
+  response.status(401).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', middleware.userExtractor, async (request, response) => {
   const body = request.body
   const blog = await Blog.findById(request.params.id)
   const newBlog = {
@@ -46,8 +60,11 @@ blogsRouter.put('/:id', async (request, response) => {
     url: blog.url,
     likes: body.likes
   }
-  await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
-  response.status(200).end()
+  if(blog.user.toString() === request.user._id.toString()){
+    await Blog.findByIdAndUpdate(request.params.id, newBlog, { new: true })
+    response.status(200).end()
+  }
+  response.status(401).end()
 })
 
 module.exports = blogsRouter
